requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
addict=2.4.0=pypi_0
appdirs=1.4.4=pypi_0
arithmetic-compressor=0.2=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py37hd23a5d3_7
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pypi_0
configargparse=1.7=pypi_0
cudatoolkit=11.6.2=hfc3e2af_13
cycler=0.11.0=pypi_0
dash=2.15.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
decorator=5.1.1=pypi_0
diff-gaussian-rasterization=0.0.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.6.1=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=4.3=hf484d3e_0
flask=2.2.5=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h267a509_2
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.3.0=h59595ed_1
gnutls=3.6.13=h85f3911_1
gridencoder=0.0.0=pypi_0
icu=73.2=h59595ed_0
idna=3.7=pyhd8ed1ab_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipython=7.34.0=pypi_0
ipywidgets=8.1.3=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h0b41bf4_3
jsonschema=4.17.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.14=h6ed2654_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libdeflate=1.14=h166bdaf_0
libffi=3.3=h58526e2_2
libgcc-ng=13.2.0=h807b86a_5
libhwloc=2.10.0=default_h2fb2949_1000
libiconv=1.17=hd590300_2
libpng=1.6.43=h2797004_0
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtiff=4.4.0=h82bc61c_5
libwebp-base=1.4.0=hd590300_0
libxcb=1.13=h7f98852_1004
libxml2=2.12.6=h232c23b_2
libzlib=1.2.13=hd590300_5
llvm-openmp=18.1.3=h4dfa4b3_0
lpips=0.1.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h8d4b97c_729
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
nbformat=5.7.0=pypi_0
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pypi_0
nettle=3.6=he412f7d_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
open3d=0.17.0=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=h7d73246_1
openssl=1.1.1w=hd590300_0
packaging=24.0=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.18.0=pypi_0
plyfile=0.8.1=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pypi_0
protobuf=4.24.4=pypi_0
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.13=haa1d7c7_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.7=2_cp37m
pytorch=1.12.1=py3.7_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytorch-scatter=2.1.0=py37_torch_1.12.0_cu116
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
retrying=1.3.4=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
sentry-sdk=1.45.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
simple-knn=0.0.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h2c6b66d_0
tbb=2021.12.0=h00ab1b0_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=1.13.1+cu116=pypi_0
torchac=0.9.3=pypi_0
torchaudio=0.13.1+cu116=pypi_0
torchvision=0.14.1+cu116=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.9.0=pypi_0
typing_extensions=4.7.1=pyha770c72_0
urllib3=2.2.1=pyhd8ed1ab_0
wandb=0.16.4=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
widgetsnbextension=4.0.11=pypi_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zipp=3.15.0=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
